Fix Fullscreen API override.
For simple background playback, neither Youtube nor Vimeo require it at the
moment.

Vimeo also generally seems to stop playback on mobile when you exit fullscreen,
but to avoid this, it is enough to simply block the corresponding event. The
Fullscreen API properties themselves need to be left alone so that the
fullscreen button in the UI still works correctly.